fixed video render
